Updated to use a better layout and structure.
This pull-request removes 99% of the code from `main.go`, instead
creating a lexer package, a parser package, an environment and an
associated interpreter package.

This gives us structure, and allows us to grow in a better way.

The initial reason for these changes were to close #1, by adding
the ability to run with a "real" parser - allowing the `if` and
`while` primitives to work in a better way.

The only major outstanding feature now is the ability to define
user-functions (via `proc`), as well as the associated bugs which
are noted in the README.md file.