Add a simple standard-library, and use it.
Updated the examples to use `{<= $a 10}` rather than the other form
(`{expr $a <= 10}`) which works now we've added corresponding helper
methods as part of a "standard library".

The library is added/loaded by the driver rather than the core
interpreter in case users don't need/want it, but this is a reasonably
easy thing to change in the future.

The only things added right now are the maths/comparision functions
and a single proc named `repeat`.

This closes #13.